Entregables/Maria del Rosario Geron Garcia/APPIUM/behave/requirements.txt
behave == 1.2.6
Appium-Python-Client==2.9.0
allure-behave==2.13.2
python-dotenv==1.0.0